deleted unneeded files, renamed probes to sensors, http request to docker is fetching ids and stats
