--- src/probes/probe2-dockerode/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "dockerode": "^3.3.4"
-  }
-}
